feature/user

+++ database/migrations/2025_04_18_224439_edit_order_id_field_in_orders_table.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Models\Celebration;
+use App\Models\Order;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -11,10 +12,13 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::table('bookings', function (Blueprint $table) {
+        Schema::table('orders', function (Blueprint $table) {
+            if (!Schema::hasColumn('orders', 'order_id')) {
+                return;
+            }
+            $table->dropForeign(['order_id']);
+            $table->dropColumn('order_id');
             $table->foreignIdFor(Celebration::class)
-                ->nullable()
-                ->after('user_id')
                 ->constrained()
                 ->cascadeOnDelete();
         });
@@ -25,9 +29,15 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::table('bookings', function (Blueprint $table) {
+        Schema::table('orders', function (Blueprint $table) {
+            if (!Schema::hasColumn('orders', 'order_id')) {
+                return;
+            }
             $table->dropForeign(['celebration_id']);
             $table->dropColumn('celebration_id');
+            $table->foreignIdFor(Order::class)
+                ->constrained()
+                ->cascadeOnDelete();
         });
     }
 };
